GUA-440: Allow non-GDS traffic to access integration
The integration environment is used by service teams outside GDS to test
their applications when they're setting our product. We need people on
those teams to be able to access the account management app.

Note this change does not make the integration environment available to
everyone on the internet. To do anything in the account management app
you first need to sign into an account. If you're not signed in you'll
get redirected to the sign in journey before seeing anything in account
management.

The sign in journey in integration is behind HTTP basic auth, so users
need to have credentials for that before they can return to account
management.

With this change I've renamed the conditions and resource names to
reference 'internal' and 'external', as that better describes the
differences between environments.