GUA-440: Allow traffic unless it's blocked by a rule
These rules have been blocking all traffic in integration since we
removed the IP allow list in #16.

Change the default action for integration and production to allow
traffic. This means all traffic is now allowed, unless it matches
one of the two rule sets.

This is the configuration we want for a public website. From the WAF
docs [[1]]:

"Allow all requests except the ones that you specify – This
is useful when you want Amazon CloudFront, Amazon API Gateway,
Application Load Balancer, AWS AppSync, or Amazon Cognito to serve
content for a public website, but you also want to block requests
from attackers."

[1]: https://docs.aws.amazon.com/waf/latest/developerguide/what-is-aws-waf.html