OLH-1991: Allow traffic from inside our build VPC
The container for the frontend post-deploy tests needs to be able to
make requests to the frontend app in our build environment. We can run
the test container inside the VPC so its external IP will be those
allocated to our NAT Gateways.

Add the external IPs of these gateways to the WAF allowlist.