OLH-2569: Remove custom WAF rules and supporting infra
We've migrated all our environments to use the security team's centrally
managed Firewall Manager which now sets all the ACL definitions.

This means we can remove the ACLs we have defined and the supporting
infrastructure like logs. This'll be a big benefit to us as the
templates in this repo are deployed manually and this was the last thing
we need to update on even a semi-regular basis.