OLH-2569: Update deployment instructions to use AWS SSO
These were written before we switched to using AWS Control Tower to
manage our AWS access and new joiners since then would not have been
able to run the commands as shown to deploy or update these templates.

The example commands were also written as a mixture of ones to run from
the repository root and from the template sub-folders. I've switched
them all to be run from inside the folders to reduce the chance of a
copy/paste error meaning we deploy the wrong template.