OLH-2838: Deploy the backup notifications stack
Add Terraform config and instructions to deploy the backup as a service
notifications stack.

We want to deploy this only to staging, integration and production
accounts but it'll be easy to add configuration for dev and build if we
need to.

The state for these resources is stored in an s3 bucket in each account,
hence the need for the per-account backend configuration. These buckets
were created as part of the AWS account bootstrap process and already
exist in all our accounts.