OLH-2842: Add a variable for environment name
We only want to deploy the restore testing stack in staging, so
Terraform needs to know which environment it's operating in.

I've deliberately only allowed staging, integration and production even
though we have more environments as these are the only ones which are
configured here. If we do need to run this Terraform in eg. dev or build
we can add those environments to the validation at the same time as we
configure them.