OLH-2976: Configure a new VPC stack in each environment
We initailly set up our VPCs using a very old version of the template
with 2 AZs. We want to use the full 3 AZs in eu-west-2, but the current
version of the template is unable to make that update simply by changing
the parameter.

We can't delete or rename subnets that have resources deployed in them
which has left us in a bit of a pickle as we also can't take those
reources down - that'll bring our apps down.

The solution is to create new VPCs first and then migrate all our
traffic over to them. This has the added benefit of moving our parameter
values into our IaC, which has been a source of problems before.