OLH-2976: Remove TF managed VPC stack
After lots of experimentation, we've decided the benefit of moving to
this new VPC isn't worth the effort of migration. I've moved all the
resources back to the old VPC so we're safe to delete it and save on
the network firewall costs. We'll revisit this work when it's time to
migrate to the spoke transit gateway VPCs.